split spec
