+++ .rspec
@@ -1,0 +1,2 @@
+--require spec_helper
+--color
