+++ bin/booker
@@ -3,7 +3,6 @@
 require_relative "../lib/booker"
 
 # booker by jeremy warner
-# open up a website or search in google chrome from cli
-# also show and/or open up google chrome bookmarks
+# open up a website or search from the cli, or open one of your bookmarks
 
-Booker.new(ARGV)
+Booker::CLI.new(ARGV)
