+++ lib/booker/version.rb
@@ -1,0 +1,3 @@
+module Booker
+  VERSION = "2.0.0"
+end
