--- lib/consts.rb
@@ -1,9 +1,0 @@
-# file for large constant strings
-
-# todo - wrap in module
-
-DEF_CONFIG = <<~EOS
-  ---
-  :searcher:  https://google.com/search?q=
-  :bookmarks: null
-EOS
